Fix config on non-Windows for input/windows
