+++ pkg/stanza/operator/input/windows/operator.go
@@ -16,10 +16,39 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )
 
-const operatorType = "windows_eventlog_input"
-
 func init() {
 	operator.Register(operatorType, func() operator.Builder { return NewConfig() })
+}
+
+// Build will build a windows event log operator.
+func (c *Config) Build(logger *zap.SugaredLogger) (operator.Operator, error) {
+	inputOperator, err := c.InputConfig.Build(logger)
+	if err != nil {
+		return nil, err
+	}
+
+	if c.Channel == "" {
+		return nil, fmt.Errorf("missing required `channel` field")
+	}
+
+	if c.MaxReads < 1 {
+		return nil, fmt.Errorf("the `max_reads` field must be greater than zero")
+	}
+
+	if c.StartAt != "end" && c.StartAt != "beginning" {
+		return nil, fmt.Errorf("the `start_at` field must be set to `beginning` or `end`")
+	}
+
+	return &Input{
+		InputOperator:    inputOperator,
+		buffer:           NewBuffer(),
+		channel:          c.Channel,
+		maxReads:         c.MaxReads,
+		startAt:          c.StartAt,
+		pollInterval:     c.PollInterval,
+		raw:              c.Raw,
+		excludeProviders: c.ExcludeProviders,
+	}, nil
 }
 
 // Input is an operator that creates entries using the windows event log api.
